Made some huge changes and variable name changes to camelcase:
Assets:
Added a bunch of new assets so I can make the stylistic changes and make the game fit the product brief of being stylistically towards a uni with all the png files and code bringing them in

GameScreen:
I've added an end state as a graduation cap that takes you to a victory screen which I implemented and renamed some variables for readability such as assets to gameAsesets, cause I was pulling from both assests file (if we were iterating through again, i'd only have one asset file at some point). linked it to a separate victory screen that I made

Lwjgl3Launcher:
Made it so that the game would launch to full screen so that when i hardcoded the locations for each of the buttons and items, it would remain consistent among devices whilst keeping development time within the breadth of the project

MenuAssets:
simialr to the assets folder, added a bunch of new png assets for stylistic purposed

MenuScreen:
Changed the sizing of all the buttons and the locations, and added a background image and then added text to the top stating our group name and the game name from the product brief 'Escape the Uni'

SettingsScreen:
Implemented the settings so that you can choose between wasd and arrow keys , with both buttons being functional and being big in size and bringing those assets from open game art

TutorialScreen:
Resized all of the buttons and the text adn made slight changes to the text for clarity and added and appropiate background for it

VictoryScreen:
Implemented another new screen state that you've won adn provides you a button to go back to the home

Next steps commit:
Output the score that iv'e already implemented the calculation for as being the time multiplied by pi on the end state screen and maybe a leaderboard if we have the time. maybe change the graduation cap animation to make it clears to the end?